requirements.txt
python-dotenv
SPARQLWrapper